refactor: deprecate chat AnnouncementColor enum
